worker/requirements.txt
python-dotenv
requests
flask
jsonify